pyproject.toml
[project]
name = "tmuacg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.6.4",
    "jishaku>=2.6.3",
    "python-dotenv>=1.2.1",
]
